refactor: 시험 응시 훅 도메인 기반 재구성 및 코드 리뷰 반영 (#62)
